requirements.txt
pandas
python-dotenv
Pillow